Added support for reading the LANG environment variable to determine the language
getLanguageFromEnvironment in Aura/Languages.hs defines currently supported
languages and fall backs to English if neither the command line argument is supplied
nor there is a $LANG environment variable. It uses pattern matching (serbian is
matched with ('s':'r':xs) for example) which might be dificult for non-haskellers, but
that was the most straight forward solution to xx_YY@zzz.www style language codes
with most parts being optional.