Relatively major re-organisation.
Purpose of this is to clean out 'Core' of all the specific AUR
related functionality and push it into the Aura.AUR module. Likewise
for the nascent ABS functionality and for virtual packages.

Dependencies is still heavily biased towards AUR. Plan is abstract the
filtering and stuff such that one can specify the filtering to take
place. Filters will also return the correct type of package to allow
this to continue.

There may also be a case for unifying AURPkg with AUR/PkgInfo and
ABSPkg with ABS/PkgInfo, but I haven't given this much thought.