Implement function to enable 'absSyncLocal'. This will try to synchronise only the existing abs tree on a refresh.
However, I haven't enabled this as default or alternative functionality
yet, because I think it should possibly be more clever in terms of
deciding when to forget about syncing individual packages and just
fetch the whole tree, or specific repositories. Currently, a 'local'
sync on an existing tree will build up an extremely large command.