Implement an '--allsource' flag.
This option causes makepkg to generate complete source packages
for any built packages, as if using the '--allsource' option to
makepkg directly. Source packages are stored in /var/cache/aura/src.

This could potentially be useful for issue #20, although that would
need the other side as well - spotting that an old source package
exists and decompressing it for use when rebuilding.

There's also some bits of necessary support for invoking makepkg's
'-S' flag, which creates a source tarball suitable for upload to the
AUR, but I don't see this as especially useful for AURA, so at the
moment it's not complete.