Fixes problem with renaming files across drives.
This was mentioned in issue #8, and is mostly fixed there. However,
it still occurs if part of the /var/cache/aura tree is mounted on
another drive.

This fix follows the normal behaviour unless it encounters an
unsupported operation (thrown when across drives), in which case
it resorts to copying the file instead.