Check whether a package is installed before syncing it.
This patch changes the behaviour of -M under the --absdeps option.

This fix partially addresses issue #125 by changing the behaviour
described in issue #122, and only syncing packages which are not
already installed.

This applies only to dependencies of a given package, and not when
a package is specified as the main package. This allows packages to be
reinstalled when required explicitly.