Update installation of packages
The user interface should be the same as before, but installation
now uses the Package and Repository types.

After a bit of user interaction, dependencies are calculated and the
resulting packages are split into pacman and buildable packages. The
pacman packages are filtered (with pacman) to remove up-to-date
packages and installed all at once with pacman. The buildable packages
are also filtered (but explicitly installed packages are kept) and
built one-at-a-time. In the future, moving to a libalpm backend will
eliminate this distinction and simplify the process.