Update -M commands
Installing and checking dependencies with -M now works. When installing
ABS packages, dependencies can be resolved in two ways. By default,
they are resolved using pacman packages. If `--absdeps` is specified,
they are resolved using the ABS tree and then using pacman packages.
This means that the user can specify which dependencies to install by
syncing them in the ABS tree first. All ABS dependencies are installed
even if they are up-to-date.

Dependency checking with -Md now has the same behavior. In particular,
dependency checking is affected by `--absdeps`.

Querying the ABS tree is handled with a separate `PkgInfo` type.