allow number parameter for --head and --tail
This allows one to specify a specific number of items to include in the
--head and --tail options to the search command. This change is
backwards-compatible in that the number parameter is optional. If the
flag is called in the same way as it was before this change, the number
is still assumed to be 10.

In other words, the following continues to work, showing the first 10
results:

```
$ aura -As --head needle
```

The following now works as well, showing the first 3 results:

```
$ aura -As --head=3 needle
```