Add --use-git/-g Flag
This commit adds a flag telling aura to use git clone to download
PKGBUILDs for building packages and adds a corresponding setting.

Note that this does not actually add said functionality to aura, but
only the flag to use as a baseline for possible implementations (using
either CLI or calling libgit directly).