pull out translated field sets into individual translations
This takes three translation functions that return an array of
translated strings and pulls out each of the strings into their own
translation function. Then we can get the same result by using
`sequence`, since the signature looks like this:

    sequence :: (Traversable t, Monad m) => t (m a) -> m (t a)

or, in this instance, where `m` is `(-> Language)` and `t` is `[]`:

    sequence :: [Language -> String] -> (Language -> [String])

There are good reasons for doing this. For one, it is very hard to add a
field to the list. You have to figure out where to insert it once for
_each language_. This is not only tedious, but it is error prone, too.
For example, in `infoFields`, the spanish translation list was lacking
an entry for "Popularity", which meant that it would label the
popularity field as "Descripción" and probably drop the description
field completely. And then, after you figure out the index in which to
insert it, you must place the english label as a placeholder for each
translation, which makes it hard to see what needs translating. As part
of this change, fields that were placeholders were removed so that it is
easy to see what is missing.

The other reason for this is that some fields are shared, like many of
the fields between `infoFields` and `absInfoFields`. Because the strings
were duplicated, there were some translations available in one but not
the other. The deduplicated fields in this change include translations
from both sources. There were no conflicting translations to resolve.