move language fields into separate module
Because the fields use some common names, and because we don't
explicitly want the rest of aura to use them directly, we can move them
to an isolated module that gets imported into the `Aura.Languages`
module. This does mean we will have to move the `Language` type and data
constructors to a shared module as well.