allow more flexible locale parsing
This improves aura's use of system locale detection and utilization.
First, instead of just reading from `LANG`, we are mimicking
`SETLOCALE(3)` to query the `LC_MESSAGES` category, which is used for
translations. This means we first prefer `LC_ALL` (used for debugging
purposes), then `LC_MESSAGES`, and then `LANG`. It is worth noting that
`pacman` uses this same logic for displaying its translated messages.
Some functions have been renamed to make this more clear. It is worth
noting that there _is_ a `setlocale` package for Haskell that acts as a
binding to the C function, but doing it manually seems to fit into
Aura's use of the pure `Environment` map more cleanly.

The other issue is that the `Languages` module was only matching exact
character codes for different languages. This is problematic, since
locales are of the format `language[_territory][.codeset][@modifier]`,
for example, `de_DE.UTF-8`. What we really want to do is only examine
the language portion of this locale code and match it against the
support map. We could parse the whole locale, but it doesn't seem like a
very necessary step at this point in time.