allow use of language flags
Currently, when you use any language flag with a command, such as

    aura -Ai aura --german

you get the error

    aura >>= Angegebene Kommandozeilen-Flags sind widersprüchlich!

which is the german translation of "Conflicting flags given!". This is
because `executeOpts` in the main executable module is strict about
which flags it is given for a command. Before `executeOpts` is called,
though, all settings flags, such as `--quiet` are stripped out so that
they do not cause conflicts. It appears that language flags should also
be stripped out, and doing so prevents this conflicting error.