feat(aura): support `chroot` in config
While not yet integrated into the build process, this will be used to
specify certain packages that the user wishes to be built via `pkgctl
build`. By default this is not activated for all packages as setting up
chroots can take some time, and local `makepkg`-based building is still
easier to inspect.

Still, building via chroots is probably safer overall and should be the
direction we move toward.