feat(-A): print entire dependency cycle
If a cycle is found during the determination of the build order, the
shortest path of the cycle(s) is displayed, instead of the unuseful
message "Found a cycle involving Foo".

This doesn't attempt to resolve the cycle in any way, but at least it is
a starting point for the user to investigate.