feat(aura): only install requested items will be installed within split packages
`nx` is an example of a package base that has no corresponding package
name. Instead, it has several child packages that are all built
together, but don't necessarily interdepend.

As of this commit, requesting `nxproxy` will only install it and its
inter-package dependencies, not any other members of `nx` that happened
to also be built.

Fixes #688