fix(-Ay): refresh all clones, not just installed packages
This takes longer but is more correct. `-Ap` also relies on these
clones, so if you happened to `-Ap` on something you don't have
installed, but the clone was old, you'd see an out-of-date PKGBUILD.

Further, the `git pull`s are somewhat flakey, so this commit also alters
the logic to merely warn about failed pulls, not to bail completely.
This at least lets the process finish and gives every package a chance
to pull.